integrate feeder
