+++ proto/settlus/settlement/v1alpha1/query.proto
@@ -1,6 +1,6 @@
 syntax = "proto3";
 
-package settlus.settlement;
+package settlus.settlement.v1alpha1;
 
 import "cosmos/base/v1beta1/coin.proto";
 import "gogoproto/gogo.proto";
