+++ proto/settlus/settlement/v1alpha1/tx.proto
@@ -1,5 +1,5 @@
 syntax = "proto3";
-package settlus.settlement;
+package settlus.settlement.v1alpha1;
 
 import "cosmos/base/v1beta1/coin.proto";
 import "gogoproto/gogo.proto";
