+++ x/oracle/keeper/keeper_test.go
@@ -263,7 +263,7 @@ func (suite *OracleTestSuite) TestKeeper_GetSetDeleteVote() {
 	vote := types.AggregateVote{
 		VoteData: []*types.VoteData{
 			{
-				Topic: types.OralceTopic_Block,
+				Topic: types.OralceTopic_BLOCK,
 				Data:  []string{"1:100:foobar"},
 			},
 		},
@@ -473,7 +473,7 @@ func (suite *OracleTestSuite) TestKeeper_ClearBallots() {
 				s.app.OracleKeeper.SetAggregateVote(s.ctx, types.AggregateVote{
 					VoteData: []*types.VoteData{
 						{
-							Topic: types.OralceTopic_Block,
+							Topic: types.OralceTopic_BLOCK,
 							Data:  []string{"1:100:foobar"},
 						},
 					},
