+++ x/oracle/keeper/msg_server_test.go
@@ -55,11 +55,11 @@ func (suite *OracleTestSuite) TestMsgServer_Vote() {
 	ownershipStr := []string{"1/0x1234567890abcdef/0x1234567890abcdef:0x77791"}
 	voteData := []*types.VoteData{
 		{
-			Topic: types.OralceTopic_Block,
+			Topic: types.OralceTopic_BLOCK,
 			Data:  blockStr,
 		},
 		{
-			Topic: types.OralceTopic_Ownership,
+			Topic: types.OralceTopic_OWNERSHIP,
 			Data:  ownershipStr,
 		},
 	}
@@ -85,11 +85,11 @@ func (suite *OracleTestSuite) TestMsgServer_Vote_should_be_failed_with_different
 	ownershipStr := []string{"1/0x1234567890abcdef/0x1234567890abcdef:0x77791"}
 	voteData := []*types.VoteData{
 		{
-			Topic: types.OralceTopic_Block,
+			Topic: types.OralceTopic_BLOCK,
 			Data:  blockStr,
 		},
 		{
-			Topic: types.OralceTopic_Ownership,
+			Topic: types.OralceTopic_OWNERSHIP,
 			Data:  ownershipStr,
 		},
 	}
@@ -113,11 +113,11 @@ func (suite *OracleTestSuite) TestMsgServer_Vote_should_be_failed_if_exceed_vote
 	ownershipStr := []string{"1/0x1234567890abcdef/0x1234567890abcdef:0x77791"}
 	voteData := []*types.VoteData{
 		{
-			Topic: types.OralceTopic_Block,
+			Topic: types.OralceTopic_BLOCK,
 			Data:  blockStr,
 		},
 		{
-			Topic: types.OralceTopic_Ownership,
+			Topic: types.OralceTopic_OWNERSHIP,
 			Data:  ownershipStr,
 		},
 	}
@@ -197,11 +197,11 @@ func (suite *OracleTestSuite) doPrevote(msgSvr types.MsgServer, voteData []*type
 func buildVoteData(blockStr, ownershipStr string) []*types.VoteData {
 	return []*types.VoteData{
 		{
-			Topic: types.OralceTopic_Block,
+			Topic: types.OralceTopic_BLOCK,
 			Data:  []string{blockStr},
 		},
 		{
-			Topic: types.OralceTopic_Ownership,
+			Topic: types.OralceTopic_OWNERSHIP,
 			Data:  []string{ownershipStr},
 		},
 	}
